refactor(analysis): one way to group duplicates, not three
Duplicate grouping was implemented three times: DuplicateGroup and VideoHash
in analysis/duplicates.py, ClusterManager and DuplicateCluster in
analysis/pipeline.py, and _select_cluster_representative in
thumbnail_clustering.py. Only the third one runs — the pipeline calls
deduplicate_by_thumbnails and nothing else.

The other two were imported by their own tests and by nothing in the
product. They passed CI, appeared in the analysis package's public surface,
and read as live to anyone grepping.

That is how a real defect stayed invisible. duplicates._compute_quality_score
weights resolution .4, bitrate .3, duration .2 and stability .1, and reads
stability from a similarity_scores dict under the key f"{asset_id}_stability".
Nothing ever wrote that key. The dict it reads is typed dict[str, int] and
holds hamming distances under plain asset ids. So the stability term was a
constant 0.5 for every video — in a scorer no memory has ever been built
with.

vulture had already found it. compute_stability_score was in
vulture-whitelist.py as "unused function", along with best_video,
other_videos and three ClusterManager methods. Eight entries recording that
roughly 850 lines were dead, silenced rather than investigated.

One thing in duplicates.py was alive: thumbnail_clustering imported
_union_find_groups from it, function-level, private name. That single import
kept the whole file plausible. It moves to thumbnail_clustering, which is its
only caller.

Deleting ClusterManager then orphaned VideoAnalysisCache.get_all_hashes,
whose only production caller was build_clusters. It had its own passing test
class, which is what made it look used. Gone too.

Also fixes the package surface. __init__ mapped ThumbnailCluster,
cluster_thumbnails and compute_thumbnail_hash to analysis.duplicates, which
never defined them: `from immich_memories.analysis import cluster_thumbnails`
raised AttributeError. A string map is invisible to mypy and to any test that
imports directly, so a test now walks __all__ and resolves every name.
