+++ vulture-whitelist.py
@@ -9,16 +9,8 @@ _.right_eye  # unused attribute (src/immich_memories/analysis/apple_vision.py:23
 _.mouth  # unused attribute (src/immich_memories/analysis/apple_vision.py:249)
 used_overrun  # unused variable (src/immich_memories/analysis/clip_refiner.py:145)
 get_content_analyzer_from_config  # unused function (src/immich_memories/analysis/content_analyzer.py:95)
-hash_frames  # unused variable (src/immich_memories/analysis/duplicates.py:32)
-computed_at  # unused variable (src/immich_memories/analysis/duplicates.py:33)
-_.best_video  # unused property (src/immich_memories/analysis/duplicates.py:98)
-_.other_videos  # unused property (src/immich_memories/analysis/duplicates.py:108)
-compute_stability_score  # unused function (src/immich_memories/analysis/duplicates.py:231)
 CONTENT_ANALYSIS_PROMPT  # unused variable (src/immich_memories/analysis/llm_response_parser.py:82)
 _.reset_session_stats  # unused method (src/immich_memories/analysis/llm_response_parser.py:148)
-_.get_duplicates  # unused method (src/immich_memories/analysis/pipeline.py:30)
-_.build_clusters  # unused method (src/immich_memories/analysis/pipeline.py:75)
-_.get_excluded_ids  # unused method (src/immich_memories/analysis/pipeline.py:169)
 _.bind_legacy_analyzer  # unused method (src/immich_memories/analysis/preview_builder.py:56)
 _.download_clip_video  # unused method (src/immich_memories/analysis/preview_builder.py:122)
 _.operational_event  # unused property (src/immich_memories/analysis/progress.py:106)
